WIP|Add check whether feedit will work in frontend
Due to inner technical workings of TYPO3, we will check whether feedit
will work at all.
If not opened from TYPO3 backend, there is no opener context.
This would prevent a working backend, which is opened by feedit.

Therefore it is checked and all features are disabled with a hint.
Additionally an explanation is displayed in admin panel.

This is WIP:

Not all buttons need to be disabled. Maybe we can / should add a css
class as selector to disable only buttons which will open the backend.